create extensions file
