+++ Shukatsu Tracker/View/ProfileImageView.swift
@@ -40,14 +40,3 @@ class ProfileImageView: UIImageView {
 }
 
 
-extension UIImageView {
-    func handleTap(gestureRecognizer: UIGestureRecognizer) {
-
-        transform = CGAffineTransform(scaleX: 0.6, y: 0.6)
-//        alpha = 0.75
-        UIView.animate(withDuration: 0.5) {
-            self.transform = CGAffineTransform.identity
-//            self.alpha = 1.0
-        }
-    }
-}
